connection: check that strings are valid UTF-8
This is required by the specification since
2bbd80c8df12 ("doc: Require strings to be UTF-8").  This commit enforces
UTF-8 validity in the implementation as well.  There are Wayland
implementations that already check for UTF-8 validity, so a client or
server that sends invalid UTF-8 in a string is already broken in
practice as well as in theory.

It is possible for libwayland users to validate UTF-8 themselves, but
this is far more error-prone than validating it during unmarshalling.
Furthemore, users may assume that since the wire protocol requires valid
UTF-8, libwayland checks for valid UTF-8.  If the user passed these
strings without validation to an API that assumes valid UTF-8 without
any checks, undefined behavior could result.

Fixes: 2bbd80c8df12 ("doc: Require strings to be UTF-8")
Signed-off-by: Demi Marie Obenour <demi@invisiblethingslab.com>